Rename CoreMedia RichText Plugin
Renamed the plugin, so that it does not signal the technology
(DataProcessor) but the purpose instead. Thus, the plugin now
serves to support CoreMedia RichText.

Also removed a version reference to the RichText version.
Assumption is, that we will transparently support possibly
new CoreMedia RichText versions in the very same plugin.