First sketch for RichText Schema Validation
Prior to writing CoreMedia RichText to server
we need to ensure, that the RichText is valid.
This especially means removing any invalid
elements we did not filter yet.

This commit contains a first sketch with the
general idea to be as close to the DTD definition
as possible.

For reference also added the (automatically
generated) Relax NG scheme, which was chosen
especially for naming the elements of the
TypeScript-based schema definition.