Documenting getData-Empty-Behavior
Documenting the options trim: 'empty' (default) and trim: 'none'
for 'CKEditor.getData()'. By default (empty), the data processor
will not be triggered, but an empty string will be saved instead.

If we want to simulate CoreMedia RichText 1.0 XML, which is
regarded empty, the option trim: 'none' has to be set.