strike = s = del # coremedia-richtext
In contrast to CKEditor 4, CKEditor 5 renders strikethrough
elements not as <strike> but as <s>. Enhanced the data processor
so that it understands both elements, and in addition to that
also the <del> element.

As we don't have a multi-key concept, extracted the common part
to a constant, which is now re-used for all through elements.