Extract Data Processing for Performance Measures
To be able to measure the performance, we need to extract
the core Data Processing part of RichTextDataProcessor, as
otherwise CKEditor API will moan on "duplicate instantiation".

The new test defaults to run only for a simple scenario, but
may easily be adapted, so that it runs with more complex data.