Fixes Deadlock on LinkTarget-Removal for "No LinkTargets"
When a link had no link-targets, the previous code run into
a deadlock, as a remove operation is only added, if there
is an attribute to remove.

Now, we check the writer.batch, if any more operations got
added and otherwise just return false to signal, that we
did not change anything.