Simplify Check for "already removed linkTarget"
Investigating in API behavior of `removeAttribute` we
found that it does not add more operations if there is
no more attribute to remove (even from previous post-fix
runs).

This code could be simplified even more, if we had
ckeditor/ckeditor5#9627. The current code uses the
suggested workaround as described in the issue.