LinkTargetCommand: Can partially set target in model
We now are able to partially set the target in the
model, but only if linkHref also got changed.

And: We cannot handle yet the text-insertion triggered
by LinkCommand if the selection is collapsed (and not
within a link). In this case, LinkCommand inserts a
text with linkHref as attributes.