LinkTarget: Be aware of createText events
`LinkCommand` comes with a feature, that if the
selection is collapsed (i.e. no text is selected)
the link will be directly added as text. In this case
we get a `DiffItemInsert` which unfortunately does not
come with neither information on the inserted node
nor with information about its attributes.

Having this, we need to take some more effort, to see
if the `DiffItemInsert` is related to the insertion
of a `linkHref` attribute.

Remaining task: If the `linkHref` was not changed,
`writer.setAttribute("linkHref", ...)` does not add
any operation we can listen to as diffItem. This
requires some extra effort.