Fixes Handling of Clearing the linkHref Attribute
Clearing a `linkHref` attribute for a text which is annotated with
`linkTarget` as well, caused an infinite recursion in pre-fixer of
`LinkTargetCommand`.

This had been caused by the less strict check in `LinkTargetModelView`
which assumed, that setting `linkHref` to empty (instead of null or
undefined) is the same as deleting `linkHref`.

As it turns out, `LinkCommand` is meant to provide the ability to set
`linkHref` to an empty string, while `UnlinkCommand` will set the
attribute value to `null` instead.

Made the check in `LinkTargetModelView` post-fixer more strict.

While doing so, prepared for the last challenge, which is to respond
to changes to `linkTarget` when `linkHref` had not been changed.