Copy & Paste LinkCommand for LinkTargetCommand
We now behave very similar to the LinkCommand on execute. Nevertheless,
as expected, this state creates two entries in undo history instead of
one, if we for example create a new link with a collapsed selection.

This is a clear vote, that setting/adjusting the attributes should
go into the post-fixer while on execution we just will store the
parameters (ranges et al.) we want to set.