Data Processor: Fixes Link Handling
Prior to this change, rules, which mapped elements to themselves got triggered
twice. This is because new elements had to be created with a new namespace, which
as side-effect triggered filtering again.

If you had a mapping rule for `a` which maps "href" to "xlink:href" this rule
got called again with the new replaced element... which already got its attributes
replaced from "href" to "xlink:href". As rules are not aware of this duplicate
call (or they should not need to deal with it), the rule now failed because the
expected "href" attribute does not exist anymore.

The behavior is now, that if an element is replaced, processing will just
continue "as if it did not change", i.e. children will be processed and it
will continue with the next sibling once, children got processed.