Fixes Setting Links on Expanded Selections
As it seems, we need to clone the ranges to iterate over them.
In addition to that, we must not reset next once LinkCommand
got executed, as this may be too early for changes to apply,
which is especially true, when only the target and not the
href-attribute got changed.