Fixes linkTarget for no linkHref update
If the LinkCommand did not trigger a change for linkHref
(or its decorators) no change had been triggered we
could listen to, to apply our changes for linkTarget.

We now listen to the linkCommand execution, to trigger
linkTarget update in case it did not get triggered
previously via post-fixer.

Using both ways (post-fixers and "after-execute") guarantees
that:

* linkTarget get set
* and there is always only one undo-entry on any change

If we only would follow one of those paths, one of the
requirements above will be broken.