Fixes Undo for Collapsed Selections
If you have a collapsed selection, which has no `linkHref`
attribute yet, `LinkCommand` will create a text node with
the `linkHref` being the text as well as the attribute.

In the previous solution we added `linkTarget` as selection
attribute, so that `LinkCommand` would respect it. This
added a history entry, though. The visible effect:

If you undo the added link-text, the selection will still
contain the `linkTarget` attribute and if you start typing,
this attribute will be added to your text.

Now we listen to the `insertContent` event instead, and, just
as before, record the target and range to set which will
then be recognized within the post-fixer.