Skip Strict URI Checking
Prior to this change we denied any URI in strict mode, which
did not match the expression as given in

https://stackoverflow.com/questions/5717093/check-if-a-javascript-string-is-a-url

Unfortunately, we observed, that CKEditor's `LinkCommand` may create
`href` attributes, which are not valid according to this pattern (for
example if using umlauts within `href`). Because of this, and because
the CoreMedia DTD does no strict checking to the URI pattern anyway,
we removed this check from `RichTextSchema` as well.