fix(dataprocessor): Fixes handling of link-targets for special cases
We needed to handle more special cases especially for `xlink:show` and
`xlink:role` attribute combinations, which are valid from CoreMedia
RichText 1.0 DTD perspective.

The new approach tries to encode any combination with the best effort to
keep data and view consistent. It introduces for example a mapping for
`xlink:show="embed" xlink:role="something"`. In previous implementations
(also applies to CKEditor 4 behavior in CoreMedia Studio) we ignored
`xlink:role` as it is unexpected for having `xlink:show` different to
`other`.

Now, we keep even this unusual rule. It gets encoded by adding an
additional underscore `_` to the target followed by the denoted
`role` attribute. Thus, for the given example we get the target
`_embed_something`.

Additional Changes
------------------

### Replaced `xlink:show="replace"` by "_self" rather than "_top"

In previous approach we decided mapping `xlink:show="replace"` to
`target="_top"`. This does not match the behavior as it is modelled
in CoreMedia Studio, where `replace` is assumed to be transformed
to `_self`.

In the long run, this should be configurable. For now, we stick to
the mapping as expected within CoreMedia Studio.

### Reduced Complexity in CoreMediaRichTextConfig

We moved some special handlers to extra files to keep
`CoreMediaRichTextConfig` smaller. This is especially useful for
anchor handlers, as we expect providing mappings for even more
attributes.

Package: ckeditor5-coremedia-richtext