test(dataprocessor): Added one Table TH/TD Use-Case
Ensure, that subsequent td's following a th are correctly transformed.

This was meant as regression test, as we observed, that CKEditor
rendered subsequent td's as th instead. But, as it seems, this
decision is done deep within the CKEditor Table Plugin.

The plugin seems to require `thead` and `tbody` sections and only
rows in `thead` provide an enabled toggle-button that this is
a heading row. If you toggle this flag at any tr for example at the
end, all rows of the table will be moved into the thead section.

It may be required to update dataprocessing accordingly.