refactor(dataprocessor): Rename `wrap` to `proxy`
`proxy` makes more clear, what this method is about. And, it does not
clash with a possible new method `wrap` to introduce, which would
wrap an existing element into another. While the latter one was
planned, it did not made it - nevertheless, it makes sense to keep
the renamed method.