fix(dataprocessor): Fixes thead/tbody in data processing
Because of ckeditor/ckeditor5#9360 we must carefully preserve any
`<thead>` elements when we want to support `<th>` elements. As of
CKEditor 5, 24.x, any `<th>` outside of `<thead>` will trigger a
whole column as having only `<th>` instead of `<td>`.

With this commit, we will mark rows, which should be restored as
part of `<thead>` with a special class `tr--header`. To streamline
the behavior, the previous class for `<th>` is now changed from
`td--heading` to `tr--header`.

We also support `<tfoot>` now. As CKEditor 5, 24.x, does not support
it, it will be merged into `<tbody>` later on.