docs(dataprocessor): Add Warning Notice to NodeProxy#delegate
In general, we should not manipulate the DOM directly in rules based on
`NodeProxy#delegate`.

For tables, we just made an exception to this. Other approaches trying
to handle a required thead in view, while forbidding it in data would
have required a much more complex API.

This exception is harmless, as only the children of a table are
manipulated. During processing this means, that any rule will see
these children on continued processing and thus, may change them.