fix(dataprocessor): Fix _self vs. _top
We decided to represent _Open in Current Tab_ as `_top` within CKEditor
view. It was `_self` before, similar to the current decision on delivery
side in CoreMedia Blueprint.

`_top` seems to be better, as it "escapes" any frames and grants, that
a link is really opened in current tab, while `_self` only opens in
current context (which may mean, to open the link within an iframe
of a given page).

Adjusted tests, data-processing and examples accordingly.

Also added the expected UI-state to the examples, which may be loaded
in example app to ease manual testing.