fix(link-plugin): Revert Decision: replace is now _self
It is not easy to decide, if `xlink:show` should be represented as
`_top` or rather as `_self`. While the specification of XLink states,
that `replace` is similar to `target=_self` `_top` may be the better
option in some cases.

Nevertheless, we eventually decided to stick to the mapping
`_self` = `replace`. This is done by this commit. In addition to that,
all "definitions" in tests and examples have been updated in a way,
that, if we ever decide different, it is just a few places to adapt.

See-also: https://www.w3.org/TR/xlink11/#show-att
See-also: https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/xlink:show