test(content-links): First minor tests for MockContentDisplayService
First tests for MockContentDisplayService for the "unmatched" case.
Unmatched means, that the regular expression did not match, which will
trigger some defined behavior. In this case, a default state shall be
represented.

Decided, to widen the support for folder/document a little bit, so that,
just as in CMS, an uneven number represents a folder, while an even
number represents a document. Also added this support for the unmatched
scenario, so that the overall state is more familiar to CMS users.