fix(content-links): Fixes Test for MockContentDisplayService
One test had the wrong expectations. The other one had to deal with
`combineLatest` which may ignore values from sub-observables, which
"have been sent in the past".

A simple `setTimeout` does the trick for now. But, we should not rely
on this pattern too much, as it bears the risk of flaky tests.