Support drag and drop of an URL.
The URL must be stored in the drag dataTransfer.data in text/uri-list, otherwise it won't be recognized.
This is a shortcome of the current implementation. We can still analyze if text/plain is an url by pattern matching of well known patterns (http/https/www.*.*/..). As far as I remember the ckeditor implementation has something like this and we can maybe make use of this.