feat(content-links): First working Draft for collapsed selection handling
Provides a first working draft for collapsed selection handling, in
case no link exists yet. In this case, CKEditor's link feature writes
the URL to the text.

In case of contents we want to override this behavior and write the
content name instead. This draft contains a first sketch of the
solution, while using a mock for the content name for now.