feat(content-links): Workaround for Insufficient Differ Information
Introduced a first sketch for a workaround of ckeditor/ckeditor5#10335,
which is, that the differ does not provide enough information to
replace a just inserted text based in the DiffInsertItem.

The current state still contains several debugging statements and it
is not capable of dealing with merged nodes: If you add a link with
the same destination (`linkHref`) as the previous text, both text nodes
will be merged into one. In this case, we cannot replace the text node,
as it got detached from DOM meanwhile, but we have to deal with the
merged text-node, which is referenced in DiffInsertItem.