feat(content-links): Working Draft for Content Name Replacement
This commit contains a working draft for content-URI replacement by
content-name for collapsed non-linked selections. As the text recorded
on `insertContent` may be detached from DOM meanwhile, because it got
merged with another node, we have to evaluate the text from the
Diff(Insert)Item instead.

This state contains some debug log statements which will be removed and
it still contains `insertContent` tracking, which may not be required
anymore. Further refactorings expected.