feat(content-links): Replace "raw content URI" by name
This commit completes the previous PoC. We are now able to replace a
raw content URI by its name when `LinkCommand` decided to insert the URI
into the text for a collapsed selection.

The implementation is now purely based on the differ, instead of
listening to `insertContent` in parallel. This is, because the data
for `insertContent` may not be valid anymore, once we reach the
post-fixer. For example the inserted text node may have been removed
meanwhile because it got merged with some other text before or after.