fix(content-links): Fix Droppable Type Guessing
Droppable type guessing now just uses the last digit of a content-ID to
decide, if a document (thus, no folder) is droppable or not.

The previous approach made content IDs, which are dividable by 4
"non-droppable documents". This clashed with the semantics introduced
by MockContentDisplayService, which only took the last digit as
"type identifier". Other digits (2nd to 4th place from the end)
are used to define content-states, unreadable states and a name.