refactor(link-target): Better "keep defaults" for `parseLinkTargetConfig`
To not require extending `parseLinkTargetConfig` for any attribute, we
possibly add to the configuration of a target in the future, enhanced
the "merge" algorithm.

This deletes the option to get the current `iconCls` attribute parsed.
This is by intention, as we plan to remove `iconCls` and replace it
by a SVG reference anyway, as it is used for image-style plugin.