feat(app): Add CodeBlock Plugin
As first approach supporting a possible `<pre>` element, added an
example to use the CodeBlock plugin. Different to CoreMedia RichText 1.0,
this approach won't support different contents for `<pre>` than `<span>`
(which will be transformed to `<code>` in view given a specific class.

Note, that this is just an example, how we may support such plugins. It
is insufficient yet, to fulfill the requirements of CoreMedia
RichText 1.0 DTD.