feat(app): Use class based alignment
Similar to CKEditor 4 in CMCC 10 and before, we use class-based
alignment (using classes `align--left` et al.).

Note, that this approach fails for headings in the current state of
data-processing. As headings are transformed having only one explicit
class, any classes applied by the alignment plugin, will be removed
on data-processing. This has to be fixed later on.