docs(pages): Experiments on Jekyll Failures
Jekyll seems to have problems mapping some links from .md to .html on
GitHub pages. Trying to identify the possible cause by further
experiments.

In Link Plugin, a link like [Test](./TEST.md) works. Thus, we now try
several ways to represent a link and see, if we can identify the
scenario, which does not work on the main page.