docs(pages): Workaround Jekyll Issue for Reference Style Links
As experiments have shown, Jekyll seems to be unable to resolve reference
style links to local markdown files, such as

```
[Test][]
[Test]: <./TEST.md>
```

This should be mapped to `TEST.html` on GitHub pages, but instead creates
a link to `TEST.md` which loads as plain-text into the browser.

The following is a workaround for just reported jekyll/jekyll#8828:

```
[Test](./TEST.md)
```

which is correctly mapped to `./TEST.html`.

See-also: jekyll/jekyll#8828