Get more specific to our content with contentId and property.
We are very flexible as long as the data processing converts to the actual data.
This means we can store in the view (as part of the image tag) almost everything we want/need to make our way to the existing data processing.
This may help to get an easier implementation as in the ckeditor 4 implementation.