feat(grs): Enable Headings h1 to h6
While h1/h6 are not part of the _raw_ data represented as
CoreMedia RichText 1.0, they are part of the data layer in
CKEditor after data-processing has been applied. In this
layer, the data representation is in HTML - which again is
different from HTML in view model.

Thus, they have to be part of the GRS configuration.
Added a corresponding comment to clarify this.