feat(grs): Fixes <u>, <s>, <strike>, et al.
It is important to learn, that we have two data representations.
One is CoreMedia RichText and the other one is the data-processed
CoreMedia RichText, which is then parsed by CKEditor into its model.

The GHS configuration must take care of this: It must apply to the
data-processed RichText rather than the raw RichText as it comes from
server. That's why we have to add for example `<strike>` to the
configured elements for GHS, while it is not part of CoreMedia RichText.

As `<strike>` is represented as `<span class="strike">`, `<strike>`
shares the same configuration as `<span>`.