fix(data-processor): Validate Tokens in ElementProxy.classList
Using invalid tokens in classList on Element from DOM will raise
errors. While these are different in detail (e.g., Chrome uses
different errors than JEST DOM) the failure behavior is the same.

Thus, decided to simulate the same behavior, rather than trying to
auto-heal certain scenarios.

For validation purpose, extended the tests to always compare to the
behavior of the JEST DOM, assuming, that it is close to for example
real-browser behavior.