feat(data-processor): Basic Support for Images
Adding basic support for images, as it is required for GRS.

Most attributes of images from CoreMedia RichText are transformed
to corresponding data attributes back and forth and are also
registered in GRS.

The `src` attribute maps to a dummy image, currently. Later, it must
be resolved to some Blob URL to get the image Blob data from Studio
REST backend.