fix(data-processor): Fixes Image on toData Processing
False-positive detection of invalid image not having a required
attribute (xlink:href). Checked for same attribute as for `<a>`
(href) but we need to check for either `data-xlink-href` or
`xlink:href` after the corresponding XLink-Mappers have been
executed.

Decided for the latter one, as it is independent of the implementation
of Xlink-Attribute mapping.