feat(grs): Fixes Language Handling for blockquote
We did not handle the xml:lang attribute of blockquote,
which caused only the "lang" attribute to be persisted.
What was missing is the mapping configuration in
CoreMedia RichText data-processing, which handles
xml:lang/lang attributes to be mapped to lang
attribute, which then is registered at GRS as allowed
attribute for `blockquote`.