feat(grs): Changes decision on pre's xml:space
Before, we decided to just remove `xml:space` in GRS, as it is
irrelevant as fixed property. For consistence with other fixed
attributes, we preserve as well, we now added `xml:space` to the known
attributes for `<pre>` in data view.

Note, that in contrast to similar use-cases like `xlink:type` for anchor
elements, we don't need to provide a mapping to some data-attribute,
because `xml:space` is a well-known attribute in HTML.