fix(#39): Don't try decoding entities
The cause of #39 was the expectation, that CKEditor 5 may provide
entities (despite XML entities) when writing to server, which may not
be valid CoreMedia RichText 1.0. This was in line with a similar setup
we had for CKEditor 4 integration before.

Experiments show, that entities are stored decoded, thus, as UTF-8
characters instead. Trying to add extra decoding only caused breaking
must-be-encoded entities such as `&lt;` and `&gt;`.