fix(data-processor): Should accept `setData("")`
Just as an empty text in data view is transformed to an empty string,
we should accept the same string as input to `toView`. Before that,
it was required to always pass a valid RichText document (here:
representing an empty state).

Now, we skip XML parsing and directly set the empty data. This much
more matches the general approach of clearing an editor by setting
its data to an empty string.