feat(data-processor): Introducing IdHelper
For convenience when working with source editing, we want to be able
also accepting content IDs for `xlink:href` values as for images
and content links.

Note, that in contrast to the IdHelper, which is part of CoreMedia
Unified API, parsing is unable detecting legacy URIs. As formatting is
only relevant for user input into source editing, there is no need to
be able to deal with these legacy URIs.