Do not do a split if it is not the first item which will be added.
The order of doing splits is related to the order of the items. If we add 3 documents and the second one is added first and does two splits because no split happened before the undo steps are weird. When undoing step 3 and 2 the first one is rendered at the beginning of the next paragraph instead of in an own paragraph.