fix(data-processor): Adjust Data Processing on Element Replacement
Prior to this we decided to continue data-processing with next sibling.
See a2c9cd9dc3e369951d8500eec658444133dd46a3 for reference.

This was, because prior to that, we always reprocessed a node, even if
just its namespace and not its name changed. This caused links to
disappear, as namespace-change was done together with mapping `href`
to `xlink:href`. When reprocessing this element, it caused the anchor
to be removed, because of missing required `href` attribute from data
view.

The new behavior triggered another bad behavior, though: All
sub-elements were processed two times and rules for the element we just
mapped to, weren't applied.

The latter one enforced to add, for example, a locale mapper for any
element we receive from data view, when gets mapped to `<span>` with
some dedicated class. This is error-prone (easy to forget) and different
to CKEditor 4 behavior, we wanted to get close to.

The first one is even worth: It also caused for example text-nodes to
be processed twice. If you think of encoding text-nodes this could
be the processing:

data view:  `Lorem & Ipsum`
intermediate data (first iteration): `Lorem &amp; Ipsum`
data (second and last iteration): `Lorem &amp;amp; Ipsum`

The changed behavior now distinguished between just namespace rename and
a real renaming of elements.