refactor(docs): Prefer `index-doc.ts` over `doc/index.ts`
The previous approach caused an extra `doc/` folder to appear in
generated TypeDoc.

The new approach is using `index-doc.ts` within the `src/` folder.
This also spares some additional configuration required otherwise.

Also experimented with a concept for nested module paths. The current
state (see ckeditor5-logging) seems to be the best approach:

In root `index-doc.ts` we explicitly declare a namespace for
`logging`. This is now visible in generated TypeDoc/TSDoc.