fix(docs): Fix Problematic index-doc.ts Exports
For globals contained in an exported .ts file, it is perceived, as if
the globals are part of the virtual namespace and not of a nested file.

One workaround is, adding an explicit namespace on export:

    export * as Anchor from "./Anchor";

This does not work, though, when the corresponding file already contains
a default export of the same name. Skipped handling that quirk for now.