feat(studio-integration): New Start for Content Mocks
Previous content mocks used some sophisticated number magic to get a
content of a given behavior by a magic id.

The new suggested approach is not using magic, but pure configuration
to get a special behavior on contents.

While yet unused with this commit, services like the
`MockContentDisplayService` are going to use this new mocking
behavior.

The desired goal is to provide an easier to use and easier to
extend API for mocking.

Note, that blob property mocking is not matured yet with this commit.