refactor(studio-integration): Refactor MockRichtextConfigurationService
The service now relies on MockContentPlugin. Simplified the droppable
magic by just any content type, which is not "document" to be denied.

dragExamples now use the plugin API to define their own MockContent,
which is regarded _Undroppable_.

Additionally, I added some methods to validate, if an expected content
exists. This should help when we do accidental refactoring of content
IDs in our mocked contents.