test(studio-integration): Tests MockContentTypes and more...
While testing `MockContentTypes`, decided that it is more convenient,
if we introduce similar type defaults as used in CoreMedia CMS or
CoreMedia Studio.

Thus, we assume that an odd ID represents a folder, while an even
ID represents a document. And, as for Studio most document may be
linked within some RichText, they are by default marked as linkable.