fix(studio-integration): Fixes endless observe_asLink
Due to some collision between ServiceAgent and Content-Mocks, an
endless call of `observe_asLink` got triggered especially for
contents having only one state.

Root Cause: For a Service not fetched with `{ fixed: true }`, the
service agent's behavior is to recreate an observable, once it
is completed.

Thus, we must not complete the observed property changes. As this is
an artificial `complete` not relevant in production scenarios, it
does no harm, that we remove it in mocking - we are just a little closer
to the behavior in production code (thus, Studio Client).