fix(content-clipboard): dropEffect now correctly applied
CKEditor's internal processing enforces the `dropEffect` to `move` for
several scenarios (depending on the `effectAllowed`). To be able to
override the `dropEffect` we must add the `dragover` listener with
lower priority than standard CKEditor listeners.

The alternative, to call `evt.stop()` is unsuitable, as it will stop the
drop location indicator not to be displayed.

That's why also `evt.stop()` has been removed for _forbidden_ state. Now
the indicator correctly shows as `forbidden` *and* we have an indicator
where the content would drop.