fix(app): Fix Droppables/Undroppables
Content #40 missed an explicit "droppable" exclusion.

Root Folder #0 missed an explicit "droppable" inclusion.
We want to have the root-folder to be droppable, as it
shows a corner case scenario having some content with
empty name.

The root-folder case is not completely artificial, as
it is possible to configure CoreMedia Studio, so that
linking to folders within RichText is allowed.