feat(studio-integration): Extend RichtextConfigurationService
We require a new method to resolve the blob-property reference to
be used in embedded images.

Design decision is, that it is under complete control of CoreMedia
Studio, how these references will look like. In general, it is
expected that for a given URI-Path `content/42` of an image document
something like `content/42#properties.data` is returned.

Theoretically, we could just request the BLOB property to use and
create our own pattern by `${uriPath}#properties.${property}`. But this
would require extra knowledge in CKEditor workspace about the structure
of these blob-referencing URI-paths.