test: Respect Jest Playwright Config
We needed to change the testEnvironment in Jest Configuration,
so that the Jest Playwright configuration is respected.
Having this, we use the globally available `page` instead
of invoking our own browser.

Additionally, we added a "screenshot on failure" feature,
as described for Jest Playwright. Small adaption:  We also
take the browser name into account to have different screenshot
paths for different browsers.