test: Access Editor Through Wrapper
Hiding complexity from tests is a good idea. The new
`HelloClassicEditor` uses a wrapper to access the editor
provided in the loaded app.

As in Jest Playwright we cannot wait asynchronously for
a state to change, we introduced a new dependency:
`wait-for-expect`. This, quite transparently, waits for
an expectation to be fulfilled eventually.