Fixes Undo/Redo Enabling/Disabling
We must not manually disable or enable commands. Instead, we must
use forceDisabled(id) and clearForceDisabled(id), so that we do not
collide with other plugins having similar requirements.

Adapted the integration to (optional) Undo-Plugin similar as done in
CKEditor's HTML Support: We will now have an `integrations` folder,
which signals dependencies to other plugins. This will make
these integrations more visible, as well, as it should ease
maintenance, if any of the dependent plugins change.