test(Playwright): Not Ready Yet: MockContent Access
This commit contains a first sketch for MockContentPlugin access to
mock contents right from within the test. Unfortunately, we have some
issues, as with Jest 27 resolving modules does not work correctly.
This commit contains a partial revert of the workaround removed when
the CKEditor plugin project moved to Jest 28. It does not work yet,
though, which requires some workarounds like typings for our own
classes.

This has to be fixed - or we wait for Jest Playwright to support
Jest 28.