build(jest): Downgrade to Jest 27.5.1
While originally only stayed for Jest Playwright tests at 27.5.1
because of playwright-community/jest-playwright#796, it seems,
that pnpm (tested for version 7) fails (sometimes) to get
dependencies straight, when two different versions of a package
are in use.

We observed (on Windows, sometimes also on Linux), that mixing
Jest 28.x and 27.x in workspace sometimes caused Jest 28.x to
have transitive dependencies resolved to incompatible Jest 27.x
(like Jest 28.x with incompatible @jest/core 27.x) and vice
versa.

Downgrading to 27.5.1 also means, that we are again affected
by facebook/jest/issues#9771 (resolved for 28.x only), so that
we need to reintroduce the enhanced-resolve workaround for all
packages again (previously only active for Jest Playwright
tests in `itest/`.