feat(richtext): Provide Simple Difference Detection
The data-processor may map different data received from a backend to
the same data view. When processing from data-view to data, these
data may have been normalized to one representation. For example,
the RichTextDataProcessor would remove obsolete namespace
declarations such as `xmlns:xlink`, if no xlinks are represented
in document.

Nevertheless, such irrelevant changes should not raise awareness
of the backend, which (in case of CoreMedia Studio) would trigger
an auto-checkout of an involved document.

With this commit, data-processors can provide an interface, which
helps to perform a _lightweight_ check, if data are considered
different by this data-processor. If in doubt, this comparison
will signal a difference.

For example, in this state, the differ does not ignore order of
attributes but instead will respect them as difference. If this
should be ignored, too, the differ needs to be adapted
accordingly.