refactor(data-processor): Simplify Data Differ
The previous solution was unflexible and required implementing custom normalizations
and thus custom `DataDiffer`. It could also not be modified from clients of the
data-processor.

The new API just relies on normalizers, which can be added to the mixin-users, such
as the data-processors it is meant to support. Having this, users of the
RichTextDataProcessor, for example, could even add more normalizers, if they
feel, the existing ones are not enough.

Example: They may add a normalizer, which always sorts the attributes by name,
so that the difference is "attribute order agnostic".