test(symbol-on-paste): Rename (Clipboard|Window)Wrapper to (Clipboard|Window)BrowserAccessor
The Wrapper concept is meant to represent a dom element and simplify the access to the wrapped dom element.
In both classes we are not wrapping a dom element but accessing browser features (clipboard web api or the user agent variable).
To prevent confusion with the names I changed it to BrowserAccessor.

Provide documentation
Moved word-document-template. The folder `resources` was adapted from the maven world but
`test-data` makes it easier to understand what the contained files are used for.