refactor(typings): Adjust Code To Typings from DefinitelyTyped
As our local typings were not always inline with those provided
by DefinitelyTyped, adapting corresponding usages.

Some errors could not be fixed, because of issues in typings
at DefinitelyTyped, at others we misused the "power of typing"
adding typings for what does not exist - or does not fully
match the signature.

Some of those issues, got fixed, but this commit also contains
a lot of `// @ts-expect-error TODO` statements. In subsequent
changes, we should validate them to eventually get rid of them.