fix(data-processor): Use Hash Normalization
Introduces Hash-Normalization. Advantage: Less expensive
to remember already normalized data.

As we want to be able to "add additional normalizers", for
example, if we detect accidental auto-checkout in not yet
normalized situations, this hash-calculation must come
last/late.

To provide this, you may now define normalizers having
a priority to grant some "order of processing".