feat(differencing): Simplify Styling of Added Newline
Previous approach was a little challenging, as an added newline
got represented like this:

```
<xdiff:span xdiff:class="diff-html-added" xdiff:id="added-diff-0">
  <br data-cke-filler="true">
</xdiff:span>
```

This is because the simple configuration in `Differencing` did not
handle empty `xdiff:span` as empty elements. So, CKEditor decided
to add some filler element to it.

To fix, we now distinguish empty from non-empty model elements in
downcast. Having this, the new representation is the same as we
had for CKEditor 4, which makes styling much easier using the
`empty` pseudo-class:

```
<xdiff:span xdiff:class="diff-html-added" xdiff:id="added-diff-0"></xdiff:span>
```

CSS styling and README have been adapted accordingly.