fix(data-processor): Fixes Observation ckeditor/ckeditor5#11786
We observed that for the document-list feature, we get
list entries only containing text to be transformed to
the text being wrapped into:

    `<span class="ck-list-bogus-paragraph">...</span>`

In previous (now reverted) approached, we applied a workaround
to remove these spans prior to writing data back to server.

As emphasized in a comment in ckeditor/ckeditor5#11786, the
actual issue is using the `DomConverter` with incomplete
configuration options. It should (and must) set the
`renderingMode` option to `data` (we only use it for transformation
to data).