fix(differencing): Add Difference Challenge to Examples
CoreMedia's server side differencing challenges difference
highlighting in that way, that, if a word and its following
space gets a inline style applied, the space appears once
as changed and once as added. The added section only contains
the space character.

In its current state, this will result in the same augmented
markup in editing view, as if a newline got added directly
behind the word. This is, what needs to be fixed.