fix(differencing): Introduce artificial xdiff:br
To ease highlighting added, removed, etc. newlines via
CSS, we change the matching `xdiff:span` to an artificial
`xdiff:br` in views (data view and subsequently editing view).

While not required to fix the actual issue (it was sufficient
to set xdiff:span as preElement), this represents the first
result of refactoring and it seems to be a good idea to
assist difference highlighting even better here.