feat(data-processor): Add HTML Data Normalization
CoreMedia RichText 1.0 when read from server is not guaranteed
to be normalized in any way. It may contain ignorable whitespace
characters, attributes may be given in any order, and the same
applies to class-attribute-values.

To work around this, we introduce normalization of HTML
structures as provided by `html-minifier-terser`. This will
remove ignorable whitespace, etc.

Note, that `html-minifier-terser` comes with a transitive dependency
to `clean-css` which would require some browserify dependencies
such as for `require('os')`. As we don't use this (for now), it should
be safe, providing a corresponding configuration for webpack, similar
to the one as provided (with this commit) in `webpack.config.js`, which
is to disable these dependencies.