fix(app): Debugging Setting/Getting/Saving Data
This commit adds some more debugging to better understand
CKEditor's processes when setting, getting and saving data.
It is especially about the `editor.model.document.version`
which is taken into account by the Autosave-Plugin.

This may provide a better option than any data-normalization,
if we just keep track of the version.